+++ src/components/Comment.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import {connect} from 'react-redux';
-import {commentSelectorFactory} from '../selectors';
+import {commentSelectorFactory} from '../selectors/comment';
 
 class Comment extends React.Component {
     constructor(props) {
